Extra logic into custom hook - ModalController
I researched a bit to see how we can make our code a little cleaner as
we move things over to hooks, and the pattern I found was to make custom
effects as needed. This helps extract function to clean up our files.

I named the custom hook ModalController in this case because it's
essentially a controller for a modal, and couldn't really thinkg of a
better name. - File is called ModalController but exports
useModalController.

I did include a few eslint fixes in this commit as well.
