+++ src/components/crisisLine/VeteransCrisisLine.tsx
@@ -3,35 +3,15 @@ import * as React from 'react';
 import Modal from '@department-of-veterans-affairs/formation-react/Modal';
 import classNames from 'classnames';
 
+import useModalController from '../../hooks/ModalController';
+import rightArrow from '../../assets/arrow-right-white.svg';
 import VeteransCrisisLinePanel from './VeteransCrisisLinePanel';
 
 import './VeteransCrisisLine.scss';
 
-import rightArrow from '../../assets/arrow-right-white.svg';
+const VeteransCrisisLine = (): JSX.Element => {
 
-function useModalController() {
-  const [modalVisible, setModalVisible] = React.useState(false);
-
-  const handleEscape = (event: KeyboardEvent) => {
-    if (event.key === 'Escape' && modalVisible) {
-      setModalVisible(false);
-    }
-  };
-
-  React.useEffect(() => {
-    document.addEventListener("keydown", handleEscape, false);
-
-    return () => {
-      document.removeEventListener("keydown", handleEscape, false);
-    };
-  }, []);
-
-  return {modalVisible, setModalVisible};
-}
-
-const VeteransCrisisLine = () => {
-
-  const {modalVisible, setModalVisible} = useModalController();
+  const { modalVisible, setModalVisible } = useModalController();
 
   return (
     <div className={classNames(
@@ -51,7 +31,7 @@ const VeteransCrisisLine = () => {
           'vads-u-padding-right--0',
           'vads-u-width--full',
           'medium-screen:vads-u-width--auto',
-      )}>
+        )}>
         <span className={classNames('vads-u-display--flex', 'vads-u-align-items--center')}>
           <span className={classNames('va-api-crisis-line-container', 'vads-u-margin-right--1')}>
             <span className={classNames('vcl', 'va-api-vcl-logo-white')} />
